RF_14.002.05 Search by auto-fill
