+++ cypress/fixtures/pomFixtures/header.json
@@ -1,0 +1,3 @@
+{
+    "project_name": "testJob"
+}
